device: reorganize helpers.
