+++ lib/device/helpers/common.js
@@ -6,7 +6,12 @@
 'use strict';
 
 const assert = require('bsert');
+const {enforce} = assert;
 const {BufferMap} = require('buffer-map');
+const KeyRing = require('bcoin/lib/primitives/keyring');
+const HDPublicKey = require('bcoin/lib/hd/public');
+const Script = require('bcoin/lib/script/script');
+const MultisigMTX = require('bmultisig/lib/primitives/mtx');
 const {Path} = require('../../path');
 const {InputData} = require('../../inputData');
 
@@ -70,4 +75,115 @@ helpers.injectMTX = function (target, source) {
     // script behaviour.
     input.witness.fromArray(sourceInput.witness.toArray());
   }
+};
+
+/**
+ * Generate redeem script
+ * @param {Object} data
+ * @param {Network} network
+ * @returns {Script} redeem script.
+ */
+
+helpers.getRedeemScript = function getRedeemScript(data, network) {
+  assert(data.multisig, 'Can not get redeem script for non-multisig input.');
+
+  const publicKeys = [];
+
+  for (const pkinfo of data.multisig.pubkeys) {
+    const hdpub = HDPublicKey.fromBase58(pkinfo.xpub, network);
+    const pk = hdpub.derivePath(pkinfo.path.toString());
+
+    publicKeys.push(pk.publicKey);
+  }
+
+  const m = data.multisig.m;
+  const n = publicKeys.length;
+
+  const script = Script.fromMultisig(m, n, publicKeys);
+
+  return script;
+};
+
+/**
+ * Create keyring from InputData
+ * @param {InputData} inputData
+ * @param {Buffer} pubKey
+ * @param {Network} network
+ * @param {Buffer?} privKey
+ */
+
+helpers.createRing = function createRing(inputData, pubKey, network, privKey) {
+  enforce(InputData.isInputData(inputData), 'inputData', 'InputData');
+  enforce(Buffer.isBuffer(pubKey), 'pubKey', 'Buffer');
+  enforce(typeof network === 'object', 'network', 'object');
+
+  if (privKey)
+    enforce(Buffer.isBuffer(privKey), 'privKey', 'Buffer');
+
+  const {coin, witness} = inputData;
+
+  let nested = false;
+
+  if (witness && coin.getType() === 'scripthash')
+    nested = true;
+
+  let redeem = null;
+  if (inputData.multisig)
+    redeem = helpers.getRedeemScript(inputData, network);
+
+  const ring = KeyRing.fromOptions({
+    witness,
+    nested,
+    publicKey: pubKey,
+    privateKey: privKey
+  });
+
+  ring.script = redeem;
+
+  return ring;
+};
+
+/**
+ * Get transaction signatures and rings
+ * @param {MTX} mtx
+ * @param {Object} inputData
+ * @param {Network} network
+ * @returns {Object} - rings and signatures
+ */
+
+helpers.applyOtherSignatures = function applyOtherSignatures(mtx, inputData, network) {
+  const msMTX = MultisigMTX.fromMTX(mtx);
+  msMTX.view = mtx.view;
+
+  for (const [i, input] of msMTX.inputs.entries()) {
+    const poKey = input.prevout.toKey();
+    const data = inputData.get(poKey);
+
+    assert(data, `Could not get metadata for input ${poKey.toString('hex')}.`);
+
+    const {multisig, coin} = data;
+
+    // We only want to apply multisig signatures.
+    if (!multisig)
+      continue;
+
+    for (const pkinfo of multisig.pubkeys.values()) {
+      if (pkinfo === '')
+        continue;
+
+      const signature = Buffer.from(pkinfo.signature, 'hex');
+      const hdpub = HDPublicKey.fromBase58(pkinfo.xpub, network);
+      const pk = hdpub.derivePath(pkinfo.path.toString());
+
+      const ring = helpers.createRing(data, pk.publicKey, network);
+
+      msMTX.template(ring);
+      msMTX.applySignature(i, coin, ring, signature, false);
+    }
+  }
+
+  const nmtx = msMTX.toMTX();
+  nmtx.view = msMTX.view;
+
+  return nmtx;
 };
