+++ lib/device/helpers/trezor.js
@@ -1,6 +1,6 @@
 /*!
  * trezorhelpers.js - Trezor input transformations and other helpers.
- * Copyright (c) 2019, Nodari Chkuaselidze (MIT License)
+ * Copyright (c) 2020, Nodari Chkuaselidze (MIT License)
  */
 
 'use strict';
