+++ lib/device/ledger.js
@@ -11,6 +11,7 @@ const bledger = require('bledger');
 const AbstractDeviceWrapper = require('./abstract');
 const {vendors} = require('../common');
 const MTX = require('bcoin/lib/primitives/mtx');
+const CoinView = require('bcoin/lib/coins/coinview');
 const common = require('./helpers/common');
 const helpers = require('./helpers/ledger');
 
@@ -215,32 +216,36 @@ class LedgerDeviceWrapper extends AbstractDeviceWrapper {
     await this.ledgerApp.signTransaction(mtx, ledgerInputs);
     this.logger.debug('Transaction was signed.');
 
-    const nmtx = helpers.applyOtherSignatures(mtx, inputDataMap, this.network);
+    const nmtx = common.applyOtherSignatures(mtx, inputDataMap, this.network);
     return nmtx;
   }
 
   /**
    * Sign transaction and return signatures.
-   * @param {bcoin.MTX} mtx
+   * @param {bcoin.TX} tx
    * @param {Object[]} inputData
    * @returns {Buffer[]} - signatures
    */
 
-  async getSignatures(mtx, inputData) {
+  async getSignatures(tx, inputData) {
     assert(!this.destroyed, 'Device no longer available.');
     assert(this.ledgerApp, 'Ledger app not found.');
 
-    inputData = common.prepareSignOptions(inputData);
+    const inputDataMap = common.prepareSignOptions(inputData);
+    const view = new CoinView();
+
+    for (const data of inputData)
+      view.addCoin(data.coin);
 
     const ledgerInputs = helpers.createLedgerInputs(
-      mtx,
-      inputData,
+      tx,
+      inputDataMap,
       this.network
     );
 
     const result = await this.ledgerApp.getTransactionSignatures(
-      mtx,
-      mtx.view,
+      tx,
+      view,
       ledgerInputs
     );
