+++ lib/device/trezor.js
@@ -7,6 +7,7 @@
 'use strict';
 
 const assert = require('bsert');
+const {enforce} = assert;
 const Logger = require('blgr');
 const trezorConnect = require('btrezor-connect');
 const TrezorConnect = trezorConnect.default;
@@ -38,9 +39,9 @@ class TrezorDeviceWrapper extends AbstractDeviceWrapper {
     super.fromOptions(options);
     assert(options);
 
-    assert(typeof options.path === 'string', 'path must be a string.');
-    assert(typeof options.label === 'string', 'label must be a string.');
-    assert(typeof options.deviceID === 'string', 'deviceID must be a string');
+    enforce(typeof options.path === 'string', 'options.path', 'string');
+    enforce(typeof options.label === 'string', 'options.label', 'string');
+    enforce(typeof options.deviceID === 'string', 'options.deviceID', 'string');
 
     this.path = options.path;
     this.label = options.label;
